+++ frontend/src/components/Layout/Layout.js
@@ -1,17 +1,22 @@
 import React, { useState } from "react"
 import "bootstrap/dist/css/bootstrap.min.css"
-import "./layout.css"
+
 import SideBar from "./Sidebar/SideBar"
 import Content from "./Content/Content"
+import "./Layout.css"
 
-const Layout = () => {
+const Layout = props => {
   const [sidebarIsOpen, setSidebarOpen] = useState(true)
   const toggleSidebar = () => setSidebarOpen(!sidebarIsOpen)
 
   return (
     <div className="App wrapper">
       <SideBar toggle={toggleSidebar} isOpen={sidebarIsOpen} />
-      <Content toggleSidebar={toggleSidebar} sidebarIsOpen={sidebarIsOpen} />
+      <Content
+        toggleSidebar={toggleSidebar}
+        sidebarIsOpen={sidebarIsOpen}
+        {...props}
+      />
     </div>
   )
 }
