+++ frontend/src/pages/page-2.tsx
@@ -2,16 +2,15 @@
 import React from "react"
 import { PageProps, Link } from "gatsby"
 
-import Layout from "../components/layout"
 import SEO from "../components/seo"
 
 const SecondPage = (props: PageProps) => (
-  <Layout>
+  <div>
     <SEO title="Page two" />
     <h1>Hi from the second page</h1>
     <p>Welcome to page 2 ({props.path})</p>
     <Link to="/">Go back to the homepage</Link>
-  </Layout>
+  </div>
 )
 
 export default SecondPage
